feat: down version 21 -> 15

+++ src/main/java/com/nixagh/contentinput/entities/Passage.java
@@ -1,13 +1,15 @@
 package com.nixagh.contentinput.entities;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
+
+import java.math.BigInteger;
 
 
 /**
@@ -31,9 +33,9 @@ public class Passage {
     @Column(name = "name")
     private String name;
     @Column(name = "resourceid")
-    private Long resourceId;
+    private BigInteger resourceId;
     @Column(name = "productid")
-    private Long productId;
+    private BigInteger productId;
     @Column(name = "scramble")
     private boolean scramble;
